+++ src/pdi_migration/validator/__init__.py
@@ -1,4 +1,5 @@
+from pdi_migration.validator.assess import assess_source
 from pdi_migration.validator.gaps import GapReport, build_gap_report
 from pdi_migration.validator.report import MigrationReport, build_report
 
-__all__ = ["GapReport", "MigrationReport", "build_gap_report", "build_report"]
+__all__ = ["GapReport", "MigrationReport", "assess_source", "build_gap_report", "build_report"]